migrant-info-platform/backend/chatbot_service/requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
beautifulsoup4
torch
ipynbname
spacy
faiss-gpu
dotenv